+++ src/main/java/store/service/StockService.java
@@ -3,10 +3,15 @@ package store.service;
 import static store.domain.constant.StockPattern.STOCK_FORMAT;
 
 import java.util.List;
+import store.domain.billing.BillingItem;
 import store.domain.stock.Stock;
 import store.domain.vo.ProductName;
 import store.repository.StockRepository;
-
+/*
+* StockService에서 재고 관련 서비스를 모두 처리해주고 싶음
+* 먼저 재고 불러오고 해당 재고들을 사용자에게 보여주기 위한 포맷을 만드는 기능을 할당
+* 추가 적으로 재고를 업데이트 하는 기능도 이쪽으로 뺄 수 있을까?? -> 바로 적용 해봄
+* */
 public class StockService {
 
     private final StockRepository stockRepository;
@@ -46,6 +51,42 @@ public class StockService {
                     null);
             stringBuilder.append(formatted);
         }
+    }
+
+    public void updateStock(List<BillingItem> billingItems) {
+        for (BillingItem billingItem : billingItems) {
+            if (billingItem.hasPromotion()) {
+                updatePromotionStock(billingItem);
+                continue;
+            }
+            updateNormalStock(billingItem);
+        }
+    }
+
+    private void updatePromotionStock(BillingItem billingItem) {
+        if (billingItem.hasNoStock()) {
+            Long quantity = billingItem.calculateDeficitQuantity();
+            decreasePromotionStockQuantity(billingItem.getOrderProductName(), billingItem.getStockQuantity());
+            decreaseNormalStockQuantity(billingItem.getOrderProductName(), quantity);
+            return;
+        }
+        decreasePromotionStockQuantity(billingItem.getOrderProductName(), billingItem.getOrderQuantity());
+    }
+
+    private void updateNormalStock(BillingItem billingItem) {
+        if (billingItem.hasNoStock()) {
+            decreaseNormalStockQuantity(billingItem.getOrderProductName(), billingItem.getStockQuantity());
+            return;
+        }
+        decreaseNormalStockQuantity(billingItem.getOrderProductName(), billingItem.getOrderQuantity());
+    }
+
+    private void decreasePromotionStockQuantity(String name, Long quantity) {
+        stockRepository.decreaseQuantityWhereNameAndPromotionIsNotNull(name, quantity);
+    }
+
+    private void decreaseNormalStockQuantity(String name, Long quantity) {
+        stockRepository.decreaseQuantityWhereNameAndPromotionIsNull(name, quantity);
     }
 
 }
